+++ docs/docs/99-api/examples/automation_demo/more_specific_demos/demo_C.js
@@ -10,13 +10,7 @@
  *      - ~/examples/automation_demo/more_specific_demos/draw_automated.svg
  */
 
-const {
-    drawAutomated,
-    separateJSON,
-    getSize,
-    drawAutomatedOtherItems,
-    MemoryModel,
-} = require("memory-viz");
+const { drawAutomated } = require("memory-viz");
 const fs = require("fs");
 
 const WIDTH = 1300;