+++ docs/docs/99-api/examples/style_demo/nostyle_demo.js
@@ -30,38 +30,34 @@ const objs = [
         stack_frame: true,
     },
     {
-        isClass: false,
         x: 350,
         y: 350,
-        name: "list",
+        type: "list",
         id: 54,
         value: [19, 43, 28, 49],
     },
     {
-        isClass: false,
         x: 750,
         y: 500,
-        name: "str",
+        type: "str",
         id: 43,
         value: "David is cool",
     },
     {
-        isClass: false,
         x: 1050,
         y: 260,
-        name: "set",
+        type: "set",
         id: 90,
         value: [36, 49, 64],
     },
     {
-        isClass: false,
         x: 1050,
         y: 500,
-        name: "dict",
+        type: "dict",
         id: 10,
         value: { x: 81, y: 100, z: 121 },
     },
-    { isClass: false, x: 750, y: 750, name: "None", id: 13, value: "None" },
+    { x: 750, y: 750, type: "None", id: 13, value: "None" },
 ];
 
 const m = draw(objs, true, configuration);