+++ memory-viz/src/memory_model.ts
@@ -758,14 +758,18 @@ export class MemoryModel {
      *
      * @param {object[]} objects - the list of objects (including stack-frames) to be drawn.
      * Each object in 'objects' must include  the following structure:
-     * @param {boolean} objects[*].isClass - Whether a user-defined class (or a stack-frame) or a built-in
+     * @param {boolean} objects[*].isClass = false - Whether a user-defined class (or a stack-frame) or a built-in
      *                                      object will be drawn. Pass true to draw a class or a stack-frame,
      *                                      and false to draw any of the types found in the 'immutable'
-     *                                      and 'collections' constants.
+     *                                      and 'collections' constants. This has a default value of false and should
+     *                                      be manually set to true only when drawing a class or stack-frame.
      * @param {number} objects[*].x - Value for x coordinate of top left corner
      * @param {number} objects[*].y - Value for y coordinate of top left corner
-     * @param {string} objects[*].name - The type of the object to draw (if isClass===true, then this is the name of the
-     *                                  corresponding class or stackframe).
+     * @param {string} objects[*].name - The name of the class or stack frame to be drawn. Note that this attribute is only
+     *                                  applicable if the object's 'isClass' attribute is true. If no classes or stack frames
+     *                                   are being drawn, this attribute can be excluded from the input.
+     * @param {string} objects[*].type - The type of the object to be drawn. If no objects are being drawn, this attribute
+     *                                   can be excluded from the input.
      * @param {number} objects[*].id - The id value of this object. If we are to draw a StackFrame, then this MUST be 'null'.
      * @param {*} objects[*].value - The value of the object. This could be anything, from an empty string to a JS object,
      *                          which would be passed for the purpose of drawing a user-defined class object, a
@@ -828,7 +832,7 @@ export class MemoryModel {
                 const size = this.drawObject(
                     obj.x,
                     obj.y,
-                    obj.name,
+                    obj.type,
                     obj.id,
                     obj.value,
                     obj.show_indexes,